Experiment with DenseMatrix32F
I tried to see if using float matrices instead of double
would be faster. Unfortunately, this is not trivial since
most impl in EJML (like CommonOps) only support 64F.

This branch has the changes but it's incomplete (it only builds
in fast mode). I ran the mnist Train and it went from 41 to 36 seconds.
F score went up a bit... from 0.94 to 9.45. So it's good news that
accuracy didn't go down and speed is faster, but I'm not eager to
merge this branch given the amound of code we need to copy from ejml.

+++ core/src/main/java/com/basistech/ninja/Function.java
@@ -20,5 +20,5 @@
 package com.basistech.ninja;
 
 public interface Function {
-    double apply(double x);
+    float apply(float x);
 }
